File Organization
I separated the suspensions into their own python file, removed worthless html files, and added additional comments to other methods.
