+++ EsportsManagementTool/EsportsManagementTool/__init__.py
@@ -23,6 +23,7 @@ app = Flask(__name__)
 import EsportsManagementTool.exampleModule
 import EsportsManagementTool.EventNotificationManager
 import EsportsManagementTool.UpdateProfile
+import EsportsManagementTool.suspensions
 
 # Change this to your secret key (can be anything, it's for extra protection)
 app.secret_key = 'your secret key'
@@ -222,7 +223,7 @@ def cleanup_inactive_users():
     Mark users as inactive if they haven't been seen in 15 minutes
     This can be called periodically or before displaying the admin panel
     """
-    cleanup_old_invalidations()
+    EsportsManagementTool.suspensions.cleanup_old_invalidations(mysql)
     try:
         cursor = mysql.connection.cursor()
         cursor.execute("""
@@ -235,6 +236,11 @@ def cleanup_inactive_users():
         cursor.close()
     except Exception as e:
         print(f"Error cleaning up inactive users: {str(e)}")
+
+# ================================
+# REGISTER SUSPENSION ROUTES
+# ==================================
+EsportsManagementTool.suspensions.register_suspension_routes(app, mysql)
 
 # ============================================
 # ROUTES (The following was developed in part with ClaudeAI
@@ -265,128 +271,6 @@ def send_verify_email(email, token):
     - 5 Brain Cells: SU Esports MGMT Tool Team.
     '''
     mail.send(msg)
-
-"""
-Method to determine whether a user session is valid (AKA: they are not suspended)
-"""
-def check_session_validity():
-    """
-    Check if current user's session has been invalidated.
-    This runs before every request to protected routes.
-    """
-    if 'loggedin' in session and 'id' in session:
-        cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-        try:
-            # Check if this user has been invalidated since they logged in
-            cursor.execute("""
-                SELECT kicked_id, reason 
-                FROM invalidated_sessions 
-                WHERE user_id = %s 
-                AND invalidated_at > %s
-                ORDER BY invalidated_at DESC 
-                LIMIT 1
-            """, (session['id'], session.get('login_time', datetime.now())))
-
-            invalidation = cursor.fetchone()
-
-            if invalidation:
-                # Session has been invalidated - clear it
-                reason = invalidation['reason']
-                user_id = session['id']
-
-                # Clear session
-                session.clear()
-
-                # Set flash message based on reason
-                if reason == 'suspension':
-                    flash('Your account has been suspended. You have been logged out.', 'error')
-                else:
-                    flash('Your session has been terminated by an administrator.', 'error')
-
-                # Delete the invalidation record so we don't keep showing the message
-                cursor.execute("DELETE FROM invalidated_sessions WHERE user_id = %s", (user_id,))
-                mysql.connection.commit()
-
-                return redirect(url_for('login'))
-        finally:
-            cursor.close()
-
-    return None
-
-"""
-Method meant to clean invalid sessions table every 24 hours to keep it from becoming congested.
-"""
-def cleanup_old_invalidations():
-    """Remove invalidation records older than 24 hours"""
-    try:
-        cursor = mysql.connection.cursor()
-        cursor.execute("""
-            DELETE FROM invalidated_sessions 
-            WHERE invalidated_at < DATE_SUB(NOW(), INTERVAL 24 HOUR)
-        """)
-        mysql.connection.commit()
-        cursor.close()
-    except Exception as e:
-        print(f"Error cleaning up invalidations: {e}")
-
-"""
-Route designed to check session validity between pages. If a user is suspended, their next action will kick them.
-"""
-@app.before_request
-def before_request_check():
-    """Run session validity check before each request"""
-    # Skip session check for certain routes
-    excluded_routes = ['login', 'logout', 'register', 'static', 'verify_email']
-
-    if request.endpoint and request.endpoint not in excluded_routes:
-        result = check_session_validity()
-        if result:
-            return result
-
-#Function to check whether a user is suspended. If so, the user is denied the ability to log in.
-#@param - user_id is the ID of the user attempting to login.
-def check_user_suspension(user_id):
-    """
-    Helper function to check if user is suspended
-    Returns (is_suspended, suspension_info)
-    """
-    try:
-        cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-
-        cursor.execute("""
-            SELECT 
-                sus_id,
-                suspended_until,
-                reason
-            FROM user_suspensions
-            WHERE user_id = %s 
-            AND is_active = TRUE 
-            AND suspended_until > NOW()
-            ORDER BY suspended_at DESC
-            LIMIT 1
-        """, (user_id,))
-
-        suspension = cursor.fetchone()
-        cursor.close()
-
-        if suspension:
-            remaining = suspension['suspended_until'] - datetime.now()
-            days = remaining.days
-            hours = remaining.seconds // 3600
-
-            return True, {
-                'reason': suspension['reason'],
-                'suspended_until': suspension['suspended_until'].strftime('%B %d, %Y at %I:%M %p'),
-                'remaining_days': days,
-                'remaining_hours': hours
-            }
-
-        return False, None
-
-    except Exception as e:
-        print(f"Error checking suspension: {e}")
-        return False, None
-
 
 """
 API route to get event details for the Event Details Modal within dashboard.html
@@ -498,7 +382,7 @@ def login():
                         # ============================================
                         # CHECK FOR SUSPENSION BEFORE ALLOWING LOGIN
                         # ============================================
-                        is_suspended, suspension_info = check_user_suspension(account['id'])
+                        is_suspended, suspension_info = EsportsManagementTool.suspensions.check_user_suspension(mysql, account['id'])
 
                         if is_suspended:
                             # Format remaining time
@@ -690,199 +574,9 @@ def eventRegister():
             if request.headers.get('X-Requested-With') == 'XMLHttpRequest' or request.accept_mimetypes.accept_json:
                 return jsonify({'success': False, 'message': msg}), 400
 
-    # Uses the event-register html file to render the page (only for direct GET requests)
-    return render_template('event-register.html', msg=msg)
-
 import EsportsManagementTool.dashboard
 from EsportsManagementTool import game
 from EsportsManagementTool import teamCreation
-
-# =======================
-# SUSPENSION ROUTES (the following was produced in tandem with ClaudeAI)
-# ========================
-"""
-Route allowing admins to suspend users for a customizable amount of time.
-"""
-@app.route('/admin/suspend-user', methods=['POST'])
-@roles_required('admin')
-def suspend_user():
-    """Suspend a user for a specified duration"""
-    try:
-        data = request.get_json()
-        user_id = data.get('user_id')
-        duration_days = int(data.get('duration_days', 0))
-        duration_hours = int(data.get('duration_hours', 0))
-        reason = data.get('reason', 'No reason provided')
-
-        if not user_id:
-            return jsonify({'success': False, 'message': 'User ID is required'})
-
-        # CONVERT user_id to int for proper comparison
-        user_id = int(user_id)
-
-        if duration_days == 0 and duration_hours == 0:
-            return jsonify({'success': False, 'message': 'Duration must be greater than 0'})
-
-        # Prevent admins from suspending themselves
-        if user_id == session['id']:
-            return jsonify({'success': False, 'message': 'You cannot suspend yourself'})
-
-        cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-
-        # Check if user exists
-        cursor.execute("SELECT id, username FROM users WHERE id = %s", (user_id,))
-        user = cursor.fetchone()
-
-        if not user:
-            cursor.close()
-            return jsonify({'success': False, 'message': 'User not found'})
-
-        # Calculate suspension end time
-        total_hours = (duration_days * 24) + duration_hours
-        suspended_until = datetime.now() + timedelta(hours=total_hours)
-
-        # Deactivate any existing active suspensions for this user
-        cursor.execute("""
-            UPDATE user_suspensions 
-            SET is_active = FALSE 
-            WHERE user_id = %s AND is_active = TRUE
-        """, (user_id,))
-
-        # Create new suspension (sus_id will auto-increment)
-        cursor.execute("""
-            INSERT INTO user_suspensions 
-            (user_id, suspended_until, reason, suspended_by, suspended_at, is_active) 
-            VALUES (%s, %s, %s, %s, NOW(), TRUE)
-        """, (user_id, suspended_until, reason, session['id']))
-
-        # **NEW: Invalidate user's session to force logout**
-        cursor.execute("""
-            INSERT INTO invalidated_sessions (user_id, reason)
-            VALUES (%s, 'suspension')
-        """, (user_id,))
-
-        # Mark user as inactive
-        cursor.execute("""
-            UPDATE user_activity 
-            SET is_active = 0
-            WHERE userid = %s
-        """, (user_id,))
-
-        mysql.connection.commit()
-        cursor.close()
-
-        # Format duration for message
-        if duration_days > 0 and duration_hours > 0:
-            duration_text = f"{duration_days} day(s) and {duration_hours} hour(s)"
-        elif duration_days > 0:
-            duration_text = f"{duration_days} day(s)"
-        else:
-            duration_text = f"{duration_hours} hour(s)"
-
-        return jsonify({
-            'success': True,
-            'message': f'User @{user["username"]} has been suspended for {duration_text}'
-        })
-
-    except Exception as e:
-        print(f"Error suspending user: {e}")
-        import traceback
-        traceback.print_exc()
-        return jsonify({'success': False, 'message': str(e)}), 500
-
-"""
-Route allowing admins to lift suspensions for users.
-"""
-@app.route('/admin/lift-suspension', methods=['POST'])
-@roles_required('admin')
-def lift_suspension():
-    """Lift an active suspension for a user"""
-    try:
-        data = request.get_json()
-        user_id = data.get('user_id')
-
-        if not user_id:
-            return jsonify({'success': False, 'message': 'User ID is required'})
-
-        cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-
-        # Deactivate all active suspensions
-        cursor.execute("""
-            UPDATE user_suspensions 
-            SET is_active = FALSE 
-            WHERE user_id = %s AND is_active = TRUE
-        """, (user_id,))
-
-        mysql.connection.commit()
-        cursor.close()
-
-        return jsonify({
-            'success': True,
-            'message': 'Suspension has been lifted successfully'
-        })
-
-    except Exception as e:
-        print(f"Error lifting suspension: {e}")
-        return jsonify({'success': False, 'message': str(e)}), 500
-
-"""
-Route designed to find the suspension status when clicking on a user in the admin panel. Displays status if suspended.
-"""
-@app.route('/api/user/<int:user_id>/suspension-status')
-@roles_required('admin')
-def get_suspension_status(user_id):
-    """Get suspension status for a user"""
-    try:
-        cursor = mysql.connection.cursor(MySQLdb.cursors.DictCursor)
-
-        # Get active suspension
-        cursor.execute("""
-            SELECT 
-                s.sus_id,
-                s.suspended_until,
-                s.reason,
-                s.suspended_at,
-                CONCAT(u.firstname, ' ', u.lastname) as suspended_by_name
-            FROM user_suspensions s
-            LEFT JOIN users u ON s.suspended_by = u.id
-            WHERE s.user_id = %s 
-            AND s.is_active = TRUE 
-            AND s.suspended_until > NOW()
-            ORDER BY s.suspended_at DESC
-            LIMIT 1
-        """, (user_id,))
-
-        suspension = cursor.fetchone()
-        cursor.close()
-
-        if suspension:
-            # Calculate remaining time
-            remaining = suspension['suspended_until'] - datetime.now()
-            days = remaining.days
-            hours = remaining.seconds // 3600
-
-            return jsonify({
-                'success': True,
-                'is_suspended': True,
-                'suspension': {
-                    'sus_id': suspension['sus_id'],
-                    'reason': suspension['reason'],
-                    'suspended_until': suspension['suspended_until'].strftime('%B %d, %Y at %I:%M %p'),
-                    'suspended_at': suspension['suspended_at'].strftime('%B %d, %Y at %I:%M %p'),
-                    'suspended_by': suspension['suspended_by_name'],
-                    'remaining_days': days,
-                    'remaining_hours': hours
-                }
-            })
-        else:
-            return jsonify({
-                'success': True,
-                'is_suspended': False
-            })
-
-    except Exception as e:
-        print(f"Error getting suspension status: {e}")
-        return jsonify({'success': False, 'message': str(e)}), 500
 
 # This is used for debugging, It will show the app routes that are registered.
 if __name__ != '__main__':
