--- EsportsManagementTool/EsportsManagementTool/templates/calendar.html
@@ -1,118 +1,0 @@
-<!--Claude AI assisted in the creation of this html document-->
-{% extends "layout.html" %}
-
-{% block title %}Calendar{% endblock %}
-
-{% block content %}
-<link rel="stylesheet" href="{{ url_for('static', filename='calendar.css') }}">
-    <div class="calendar-wrapper">
-        <!-- Navigation and Month/Year Header -->
-        <div class="calendar-navigation">
-            <a href="{{ url_for('calendar', year=prev_year, month=prev_month) }}" class="nav-button" id="prev-month">
-                <i class="fas fa-chevron-left"></i> Previous
-            </a>
-            <h1 class="calendar-month-year">{{ month_name }} {{ year }}</h1>
-            <a href="{{ url_for('calendar', year=next_year, month=next_month) }}" class="nav-button" id="next-month">
-                Next <i class="fas fa-chevron-right"></i>
-            </a>
-        </div>
-
-        <!-- Days of week header -->
-        <div class="calendar-header">
-            <div class="calendar-day-name">SUN</div>
-            <div class="calendar-day-name">MON</div>
-            <div class="calendar-day-name">TUES</div>
-            <div class="calendar-day-name">WED</div>
-            <div class="calendar-day-name">THURS</div>
-            <div class="calendar-day-name">FRI</div>
-            <div class="calendar-day-name">SAT</div>
-        </div>
-
-        <!-- Calendar grid -->
-        <div class="calendar-grid">
-            {% for week in month_calendar %}
-                {% for day in week %}
-                    {% if day == 0 %}
-                        <!-- Empty cell for days outside this month -->
-                        <div class="calendar-cell empty"></div>
-                    {% else %}
-                        <!-- Create date string for this day -->
-                        {% set day_date = '%04d-%02d-%02d'|format(year, month, day) %}
-                        {% set is_today = (day_date == today_str) %}
-                        
-                        <!-- Calendar cell with conditional today class -->
-                        <div class="calendar-cell {% if is_today %}today{% endif %}">
-                            <div class="day-number">{{ day }}</div>
-                            
-                            <!-- Show events for this day -->
-                            <div class="events-container">
-                                {% if day_date in events_by_date %}
-                                    {% set day_events = events_by_date[day_date] %}
-                                    
-                                    <!-- Show first 3 events -->
-                                    {% for event in day_events[:3] %}
-                                        <a href="{{ url_for('event_details', event_id=event.id) }}" class="event-link">
-                                            <div class="event" title="Click to view details">
-                                                {% if event.time %}
-                                                    <strong>{{ event.time }}</strong> - 
-                                                {% endif %}
-                                                {{ event.title }}
-                                            </div>
-                                        </a>
-                                    {% endfor %}
-                                    
-                                    <!-- Show "+X more" if there are more than 3 events -->
-                                    {% if day_events|length > 3 %}
-                                        <div class="event-overflow">
-                                            +{{ day_events|length - 3 }} more
-                                        </div>
-                                    {% endif %}
-                                {% endif %}
-                            </div>
-                        </div>
-                    {% endif %}
-                {% endfor %}
-            {% endfor %}
-        </div>
-        
-        <!-- Today button to jump back to current month -->
-        <div class="calendar-footer">
-            <a href="{{ url_for('calendar') }}" class="today-button">
-                <i class="fas fa-calendar-day"></i> Jump to Today
-            </a>
-        </div>
-    </div>
-
-    <script>
-        // Quick jump to selected month/year
-        function jumpToDate() {
-            const month = document.getElementById('month-select').value;
-            const year = document.getElementById('year-select').value;
-            window.location.href = `/calendar/${year}/${month}`;
-        }
-
-        // Keyboard navigation
-        document.addEventListener('keydown', function(event) {
-            // Left arrow - previous month
-            if (event.key === 'ArrowLeft') {
-                document.getElementById('prev-month').click();
-            }
-            // Right arrow - next month
-            else if (event.key === 'ArrowRight') {
-                document.getElementById('next-month').click();
-            }
-        });
-
-        // Allow Enter key in dropdowns to jump
-        document.getElementById('month-select').addEventListener('keypress', function(event) {
-            if (event.key === 'Enter') {
-                jumpToDate();
-            }
-        });
-        document.getElementById('year-select').addEventListener('keypress', function(event) {
-            if (event.key === 'Enter') {
-                jumpToDate();
-            }
-        });
-    </script>
-{% endblock %}
