--- EsportsManagementTool/EsportsManagementTool/templates/event-details.html
@@ -1,203 +1,0 @@
-{% extends "base.html" %}
-
-{% block title %}{{ event.name }} - Event Details{% endblock %}
-
-{% block extra_css %}
-<style>
-    .event-details-wrapper {
-        max-width: 800px;
-        margin: 0 auto;
-        padding: 2rem;
-    }
-
-    .event-details-header {
-        display: flex;
-        justify-content: space-between;
-        align-items: center;
-        margin-bottom: 2rem;
-    }
-
-    .event-details-header h1 {
-        font-size: 2rem;
-        color: var(--text-primary);
-    }
-
-    .back-button {
-        display: flex;
-        align-items: center;
-        gap: 0.5rem;
-        padding: 0.75rem 1.5rem;
-        background: transparent;
-        color: var(--text-secondary);
-        text-decoration: none;
-        border: 1px solid var(--border-color);
-        border-radius: 8px;
-        font-size: 0.875rem;
-        font-weight: 500;
-        transition: all 0.2s;
-    }
-
-    .back-button:hover {
-        background: var(--hover-bg);
-        color: var(--text-primary);
-        border-color: var(--stockton-blue);
-    }
-
-    .event-details-card {
-        background: var(--card-bg);
-        border: 1px solid var(--border-color);
-        border-radius: 12px;
-        padding: 2rem;
-    }
-
-    .detail-section {
-        display: flex;
-        gap: 1rem;
-        padding: 1.5rem 0;
-        border-bottom: 1px solid var(--border-color);
-    }
-
-    .detail-section:last-child {
-        border-bottom: none;
-    }
-
-    .detail-section.full-width {
-        flex-direction: column;
-    }
-
-    .detail-icon {
-        width: 40px;
-        height: 40px;
-        background: rgba(121, 189, 233, 0.1);
-        border-radius: 8px;
-        display: flex;
-        align-items: center;
-        justify-content: center;
-        color: var(--stockton-blue);
-        flex-shrink: 0;
-    }
-
-    .detail-icon i {
-        font-size: 1.25rem;
-    }
-
-    .detail-content {
-        flex: 1;
-    }
-
-    .detail-content h3 {
-        font-size: 0.875rem;
-        font-weight: 600;
-        color: var(--text-secondary);
-        text-transform: uppercase;
-        letter-spacing: 0.05em;
-        margin-bottom: 0.5rem;
-    }
-
-    .detail-content p {
-        font-size: 1rem;
-        color: var(--text-primary);
-        line-height: 1.6;
-    }
-
-    @media (max-width: 768px) {
-        .event-details-header {
-            flex-direction: column;
-            gap: 1rem;
-            align-items: flex-start;
-        }
-
-        .event-details-header h1 {
-            font-size: 1.5rem;
-        }
-    }
-</style>
-{% endblock %}
-
-{% block content %}
-<div class="event-details-wrapper">
-    <div class="event-details-header">
-        <h1>{{ event.name }}</h1>
-        <a href="{{ url_for('dashboard') }}" class="back-button">
-            <i class="fas fa-arrow-left"></i> Back to Calendar
-        </a>
-    </div>
-
-    <div class="event-details-card">
-        <div class="detail-section">
-            <div class="detail-icon">
-                <i class="fas fa-calendar-alt"></i>
-            </div>
-            <div class="detail-content">
-                <h3>Date</h3>
-                <p>{{ event.date }}</p>
-            </div>
-        </div>
-
-        {% if event.start_time %}
-        <div class="detail-section">
-            <div class="detail-icon">
-                <i class="fas fa-clock"></i>
-            </div>
-            <div class="detail-content">
-                <h3>Time</h3>
-                <p>
-                    {{ event.start_time }}
-                    {% if event.end_time %}
-                        - {{ event.end_time }}
-                    {% endif %}
-                </p>
-            </div>
-        </div>
-        {% endif %}
-
-        {% if event.location %}
-        <div class="detail-section">
-            <div class="detail-icon">
-                <i class="fas fa-map-marker-alt"></i>
-            </div>
-            <div class="detail-content">
-                <h3>Location</h3>
-                <p>{{ event.location }}</p>
-            </div>
-        </div>
-        {% endif %}
-
-        {% if event.game %}
-        <div class="detail-section">
-            <div class="detail-icon">
-                <i class="fas fa-gamepad"></i>
-            </div>
-            <div class="detail-content">
-                <h3>Game</h3>
-                <p>{{ event.game }}</p>
-            </div>
-        </div>
-        {% endif %}
-
-        {% if event.event_type %}
-        <div class="detail-section">
-            <div class="detail-icon">
-                <i class="fas fa-tag"></i>
-            </div>
-            <div class="detail-content">
-                <h3>Event Type</h3>
-                <p>{{ event.event_type }}</p>
-            </div>
-        </div>
-        {% endif %}
-
-        <div class="detail-section full-width">
-            <div style="display: flex; gap: 1rem;">
-                <div class="detail-icon">
-                    <i class="fas fa-info-circle"></i>
-                </div>
-                <div class="detail-content">
-                    <h3>Description</h3>
-                    <p>{{ event.description }}</p>
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-{% endblock %}
