--- EsportsManagementTool/EsportsManagementTool/templates/event-register.html
@@ -1,117 +1,0 @@
-{% extends 'layout.html' %}
-
-{% block title %}Create New Event{% endblock %}
-
-{% block content %}
-<div class="create-event-container">
-    <div class="create-event-modal">
-        <h2>Create New Event</h2>
-        <p class="subtitle">Add a new event or match to the calendar</p>
-
-        <form method="post" class="event-form">
-            <label for="eventName">Event Name</label>
-            <input type="text" id="eventName" name="eventName" placeholder="Enter event name" required>
-
-            <div class="form-row">
-                <div class="form-group">
-                    <label for="eventType">Event Type</label>
-                    <select id="eventType" name="eventType" required>
-                        <option>Event</option>
-                        <option>Match</option>
-                        <option>Practice</option>
-                        <option>Tournament</option>
-                    </select>
-                </div>
-
-                <div class="form-group">
-                    <label for="game">Game (Optional)</label>
-                    <select id="game" name="game">
-                        <option selected disabled>Select game</option>
-                        <option>Apex Legends</option>
-                        <option>Call of Duty</option>
-                        <option>Counter-Strike 2</option>
-                        <option>EA FC</option>
-                        <option>Fortnite</option>
-                        <option>Game Changers</option>
-                        <option>Hearthstone</option>
-                        <option>League of Legends</option>
-                        <option>Madden NFL</option>
-                        <option>Mario Kart</option>
-                        <option>Marvel Rivals</option>
-                        <option>NBA 2K</option>
-                        <option>Overwatch 2</option>
-                        <option>Rainbow 6: Siege</option>
-                        <option>Rocket League</option>
-                        <option>Smash Ultimate</option>
-                        <option>Street Fighter 6</option>
-                        <option>Team Fight Tactics</option>
-                        <option>Valorant</option>
-                    </select>
-                </div>
-            </div>
-
-            <div class="form-row">
-                <div class="form-group">
-                    <label for="date">Date</label>
-                    <input type="date" id="date" name="eventDate" required>
-                </div>
-
-                <div class="form-group">
-                    <label for="startTime">Start Time</label>
-                    <input type="time" id="startTime" name="startTime" required>
-                </div>
-
-                <div class="form-group">
-                    <label for="endTime">End Time</label>
-                    <input type="time" id="endTime" name="endTime" required>
-                </div>
-            </div>
-
-            <label for="location">Location</label>
-            <select id="location" name="eventLocation" required>
-                <option disabled selected value="">Select location</option>
-                <option>Campus Center</option>
-                <option>Campus Center Coffee House</option>
-                <option>Campus Center Event Room</option>
-                <option>D-108</option>
-                <option>Esports Lab (Commons Building 80)</option>
-                <option>Lakeside Lodge</option>
-                <option>Online</option>
-                <option value="other">Other</option>
-            </select>
-
-            <input type="text" id="customLocation" name="customLocation" placeholder="Enter Custom location" style="display:none; margin-top:8px;">
-
-            <label for="description">Description</label>
-            <textarea id="description" name="eventDescription" placeholder="Enter event description"></textarea>
-
-            <button type="submit" class="create-btn">Create Event</button>
-
-            {% if msg %}
-                <p class="message">{{ msg }}</p>
-            {% endif %}
-        </form>
-    </div>
-</div>
-
-<!-- Script to toggle and animate "Other" location input -->
-<script>
-    document.addEventListener("DOMContentLoaded", function() {
-        const locationSelect = document.getElementById("location");
-        const customInput = document.getElementById("customLocation");
-
-        locationSelect.addEventListener("change", function() {
-            if (this.value === "other") {
-                customInput.style.display = "block";
-                customInput.required = true;
-                customInput.classList.add("fade-in");
-            } else {
-                customInput.classList.remove("fade-in");
-                customInput.style.display = "none";
-                customInput.required = false;
-                customInput.value = "";
-            }
-        });
-    });
-</script>
-{% endblock %}
